+++ DLVN-SDK/src/main/java/com/edoctor/dlvn_sdk/widget/AppointmentWidgetList.kt
@@ -2,11 +2,14 @@ package com.edoctor.dlvn_sdk.widget
 
 import android.content.Context
 import android.util.AttributeSet
+import android.util.Log
 import android.view.LayoutInflater
+import android.view.View
 import android.widget.LinearLayout
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.edoctor.dlvn_sdk.R
+import com.edoctor.dlvn_sdk.store.AppStore
 
 class AppointmentWidgetList @JvmOverloads constructor(
     context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
@@ -14,11 +17,13 @@ class AppointmentWidgetList @JvmOverloads constructor(
 
     private var recyclerView: RecyclerView
     private var activeIndicator: ScrollIndicatorView
+    private var indicatorHolder: LinearLayout
 
     init {
         LayoutInflater.from(context).inflate(R.layout.appointment_widget_list_layout, this, true)
         recyclerView = findViewById(R.id.widgetListEdr)
-        activeIndicator = findViewById(R.id.scrollIndicator)
+        activeIndicator = findViewById(R.id.scrollIndicatorEdr)
+        indicatorHolder = findViewById(R.id.indicatorHolderEdr)
 
         recyclerView.layoutManager = LinearLayoutManager(context, LinearLayoutManager.HORIZONTAL, false)
         recyclerView.adapter = AppointmentListAdapter()
@@ -39,6 +44,14 @@ class AppointmentWidgetList @JvmOverloads constructor(
                 activeIndicator.translationX = translationX
             }
         })
+
+        AppStore.updateWidgetListDisplay = {
+            when (it) {
+                "LOG_OUT" -> indicatorHolder.visibility = View.GONE
+                "HAS_DATA" -> indicatorHolder.visibility = View.VISIBLE
+                else -> indicatorHolder.visibility = View.GONE
+            }
+        }
     }
 
     private fun dpToPx(dp: Float): Int {